ajustes

+++ MediatrExample.ApplicationCore/Domain/PreparacionPasta.cs
@@ -8,8 +8,8 @@ public class PreparacionPasta
     public int PreparacionPastaId { get; set; }
     public string NombreVariable { get; set; } = default!;
     public string UnidadMedida { get; set; } = default!;
-    public int ValorMinimo { get; set; }
-    public int ValorMaximo { get; set; }
+    public decimal ValorMinimo { get; set; }
+    public decimal ValorMaximo { get; set; }
     public bool Obligatoria { get; set; }
     public bool Estado { get; set; }
 }
