+++ MediatrExample.ApplicationCore/Features/MateriaPrimaFeatures/Queries/GetMateriaPrimaQuery.cs
@@ -62,8 +62,8 @@ public class GetMateriaPrimaQueryResponse
     public string CodigoSap { get; set; } = default!;
     public string NombreVariable { get; set; } = default!;
     public string UnidadMedida { get; set; } = default!;
-    public int ValorMinimo { get; set; }
-    public int ValorMaximo { get; set; }
+    public decimal ValorMinimo { get; set; }
+    public decimal ValorMaximo { get; set; }
     public bool Obligatoria { get; set; }
     public bool Estado { get; set; }
     public List<UnidadMedida>? Unidades { get; set; }
