+++ MediatrExample.ApplicationCore/Features/PreparacionPastaFeatures/Commands/UpdatePreparacionPastaCommand.cs
@@ -13,8 +13,8 @@ public class UpdatePreparacionPastaCommand : IRequest
     public string PreparacionPastaId { get; set; } = default!;
     public string NombreVariable { get; set; } = default!;
     public string UnidadMedida { get; set; } = default!;
-    public int ValorMinimo { get; set; }
-    public int ValorMaximo { get; set; }
+    public decimal ValorMinimo { get; set; }
+    public decimal ValorMaximo { get; set; }
     public bool Obligatoria { get; set; }
     public bool Estado { get; set; }
 }
