+++ MediatrExample.ApplicationCore/Features/PreparacionPastaFeatures/Queries/GetListPreparacionPastaQuery.cs
@@ -36,8 +36,8 @@ public class GetListPreparacionPastaQueryResponse
     public string PreparacionPastaId { get; set; } = default!;
     public string NombreVariable { get; set; } = default!;
     public string UnidadMedida { get; set; } = default!;
-    public int ValorMinimo { get; set; }
-    public int ValorMaximo { get; set; }
+    public decimal ValorMinimo { get; set; }
+    public decimal ValorMaximo { get; set; }
     public bool Obligatoria { get; set; }
     public bool Estado { get; set; }
     public string StrValorMinimo { get; set; } = default!;
