+++ MediatrExample.ApplicationCore/Features/PreparacionPastaFeatures/Queries/GetPreparacionPastaQuery.cs
@@ -61,8 +61,8 @@ public class GetPreparacionPastaQueryResponse
     public string PreparacionPastaId { get; set; } = default!;
     public string NombreVariable { get; set; } = default!;
     public string UnidadMedida { get; set; } = default!;
-    public int ValorMinimo { get; set; }
-    public int ValorMaximo { get; set; }
+    public decimal ValorMinimo { get; set; }
+    public decimal ValorMaximo { get; set; }
     public bool Obligatoria { get; set; }
     public bool Estado { get; set; }
     public List<UnidadMedida>? Unidades { get; set; }
